djago_ai/django_ai/requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.3.0
category-encoders==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-imagekit==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-yasg==1.21.4
exceptiongroup==1.0.4
executing==1.1.0
firebase-admin==6.0.1
flatbuffers==23.3.3
fonttools==4.39.0
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
graphene==3.2.1
graphene-django==3.0.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
h5py==3.8.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.5.0
itypes==1.2.0
jax==0.4.5
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
keras==2.12.0rc1
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.4
numpy==1.23.5
oauthlib==3.2.2
openapi-codec==1.3.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.4.0
pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.31
proto-plus==1.22.2
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
scikit-learn==1.2.1
scipy==1.10.1
selenium==4.6.0
simplejson==3.18.1
six==1.16.0
sklearn2==0.0.13
sklearn3==0.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
statsmodels==0.13.5
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc1
tensorflow-estimator==2.12.0rc0
tensorflow-intel==2.12.0rc1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.5
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
